Cleanups * Resolved parsing on sequence map. * Removed probability node check so default can be 500 on scalar or 0 on sequence map. * Moved skills back into the main database as there is no differences between pre-renewal and renewal. * Store probability as uint16. * Corrected import file move. * Cleaned up headers.

+++ src/map/skill.hpp
@@ -372,7 +372,7 @@ extern struct s_skill_arrow_db skill_arrow_db[MAX_SKILL_ARROW_DB];
 /// Abracadabra database
 struct s_skill_abra_db {
 	uint16 skill_id; /// Skill ID
-	std::array<int, MAX_SKILL_LEVEL> per; /// Probability summoned
+	std::array<uint16, MAX_SKILL_LEVEL> per; /// Probability summoned
 };
 
 class AbraDatabase : public TypesafeYamlDatabase<uint16, s_skill_abra_db> {
